requirements.txt
fastapi==0.122.0
uvicorn==0.38.0
llama-cpp-python==0.3.16
numpy==2.3.5
